code cleaned and documented
